fix: Int64 literal for obj.size default value

+++ Sources/R2Drop/Services/R2Service.swift
@@ -68,7 +68,7 @@ final actor R2Service {
             guard let key = obj.key, let lastMod = obj.lastModified else { continue }
             let file = R2File(
                 key: key,
-                size: obj.size ?? 0,
+                size: obj.size ?? Int64(0),
                 lastModified: lastMod,
                 etag: obj.eTag ?? ""
             )
